Cache and ContentType::HTML fixes
* Cache: delete_cache_file only calls rm when file exists
* HTML: correctly process newlines in form links

git-archimport-id: e@ericrichardson.com--2004-eThreads/eThreads--eTrevolution--2.0--patch-18

+++ Object/Cache.pm
@@ -212,7 +212,10 @@ sub delete_cache_file {
 	) if ($name !~ m!^[\w\d\.]+$!);
 
 	my $file = $class->{_}->core->settings->{dir}{cache} . "/cache." . $name;
-	`rm $file`;
+
+	if (-e $file) {
+		`rm $file`;
+	}
 }
 
 #----------
